Fixed Bundling bugs

+++ Encoding.py
@@ -369,7 +369,7 @@ def ToPytorchData(par_data,bc,tol=0.0,topology=None, label_data=None,center=Fals
         
         data = Data(pos=TorchData[:,:3],x=TorchData[:,3:],edge_index=TorchTopology,mask=RealParticleMask,edge_mask=edge_mask)
         if label_data is not None:
-            data.y = torch.from_numpy(label_data)
+            data.y = torch.from_numpy(label_data.astype(float))
             
         if center == True:
             center = T.Center()
